cleaning up house before I move the module to v1 status.
