+++ AzureDevOpsMgmt/AzureDevOpsMgmt.Core/Authenticators/BarerTokenAuthenticator.cs
