+++ AzureDevOpsMgmt/AzureDevOpsMgmt.Core/AzureDevOpsMgmt.Core.csproj
