+++ AzureDevOpsMgmt/AzureDevOpsMgmt.Core/Cmdlets/Accounts/JoinAccountAndPatToken.cs
