+++ AzureDevOpsMgmt/AzureDevOpsMgmt.Core/Cmdlets/Accounts/RemoveAccount.cs
