+++ AzureDevOpsMgmt/AzureDevOpsMgmt.Core/Cmdlets/Accounts/RemovePatToken.cs
