+++ AzureDevOpsMgmt/AzureDevOpsMgmt.Core/Exceptions/ObjectExistsException.cs
