+++ AzureDevOpsMgmt/AzureDevOpsMgmt.Core/Helpers/CmdletHelpers.cs
