+++ AzureDevOpsMgmt/AzureDevOpsMgmt.Core/Helpers/JsonNetSerializer.cs
