+++ AzureDevOpsMgmt/AzureDevOpsMgmt.Core/Helpers/OperationCheck.cs
