+++ AzureDevOpsMgmt/AzureDevOpsMgmt.Core/Models/AzureDevOpsConfiguration.cs
