+++ AzureDevOpsMgmt/AzureDevOpsMgmt.Core/Models/CurrentConnection.cs
