+++ AzureDevOpsMgmt/AzureDevOpsMgmt.Core/Models/UserConfiguration.cs
