+++ AzureDevOpsMgmt/AzureDevOpsMgmt.Core/Properties/AssemblyInfo.cs
