+++ AzureDevOpsMgmt/AzureDevOpsMgmt.Core/Resources/FileNames.resx
