+++ AzureDevOpsMgmt/AzureDevOpsMgmt.Core/Resources/StaticStrings.resx
