+++ AzureDevOpsMgmt/AzureDevOpsMgmt.Core/Serialization/JsonPatchOperationContractResolver.cs
