+++ AzureDevOpsMgmt/AzureDevOpsMgmt.Core/app.config
