+++ AzureDevOpsMgmt/AzureDevOpsMgmt.Core/packages.config
